Wrote my old C code in Java instead
I remembered finding this problem fun when I tried it in C (granted, I had a bit of help with the cs50 library), so I thought I'd try it again in Java. The majority is the same, but I think I made a couple of improvements. I'll add error handling to the user input when I learn how to deal with unexpected data type entries.