Create Fizzbuzz.java
Was watching a Tom Scott video and he suggested writing this, so I did. At first all of the logic was one line of code but when he mentioned that you should really plan for the future, I agreed that was better practice and made it much easier to edit/add to in the future if needed.